+++ libs/libruntime/src/vfs/iface/server.rs
@@ -48,23 +48,23 @@ pub trait VfsServer: Send + Sync {
         &self,
         sender_id: u64,
         handle: Handle,
-        offset: u64,
+        offset: usize,
         buffer: &mut [u8],
-    ) -> Result<u64, Self::Error>;
+    ) -> Result<usize, Self::Error>;
 
     async fn write(
         &self,
         sender_id: u64,
         handle: Handle,
-        offset: u64,
+        offset: usize,
         buffer: &[u8],
-    ) -> Result<u64, Self::Error>;
+    ) -> Result<usize, Self::Error>;
 
     async fn resize(
         &self,
         sender_id: u64,
         handle: Handle,
-        new_size: u64,
+        new_size: usize,
     ) -> Result<(), Self::Error>;
 
     async fn list(
